+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- SASL token provider plugin options are now working in kubernetes mode.
+
 ## 5.8.0 - 2025-05-16
 
 ## 5.7.0 - 2025-04-02