+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- [#255](https://github.com/deviceinsight/kafkactl/pull/255) Set resource name when creating a cluster ACL
+
+### Fixed
+- SASL token provider plugin options are now working in kubernetes mode.
 
 ## 5.8.0 - 2025-05-16